Update README.md
This is a simple version of the Amazon website made using just HTML and CSS. It includes a homepage layout with a header, search bar, navigation menu, product sections, and a footer. It looks like Amazon, but it doesn’t have real shopping features (like payments or logins) — it’s just for practice or design demo purposes.